Completed MVP and Stretch...
Finished the login route for front end.
it works the same as the signup route.
also added a logout that was passed down to users page thay on click
will remove the token from local storage and push you back to the home
page.  had to add withRoute to app.js for this to work.